tests/requirements.txt
FeinCMS>=1.9
coverage
factory-boy==2.5.1
pytz
